# Hypatiamat-BackOffice/apiDados/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api Hypatiamat",
        "description": "Aqui encontra-se a documentação de toda a api de dados do Hypatiamat, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3050",
    "basePath": "/",
...
    "paths": {
...
      "/aplicacoes/turmas/{turma}":{
        "get": {
          "tags": [
            "Turmas"
          ],
          "summary": "Dado o nome da turma, devolve a informação relativa a essa turma.",
          "description": "Devolve a informação relativa a uma turma.",
          "operationId": "getTurmaById",
          "parameters": [
            {
              "name": "turma",
              "in": "path",
              "description": "Nome da turma a retornar.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Retorna a informação relativa a uma turma.",
                "schema": {
                    "$ref": "#/definitions/Turma"   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        },
        "delete":{
            "tags": [
              "Turmas"
            ],
            "summary": "Dado um identifcador da turma, pode apagá-la.",
            "description": "Apagar a informação da turma.",
            "operationId": "deleteTurmaById",
            "parameters": [
              {
                "name": "idTurma",
                "in": "path",
                "description": "Identificador da turma a apagar.",
                "required": true,
                "type": "integer"
              }
            ],
            "responses": {
              "200": {
                  "description": "A turma foi apagada."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
      },
...
    },
    "securityDefinitions":{
      "api_key":{
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "security": [{
       "api_key": []
    }],
    "definitions": {
...
        "Turma":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "idprofessor": {
              "type": "string"
            },
            "turma":{
              "type":"string"
            }
          }
        },
...
    }   
}